refactor: List<Car> 일급 콜렉션 생성
