+++ src/test/java/carracing/domain/CarRacingGameTest.java
@@ -9,19 +9,6 @@ import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
 public class CarRacingGameTest {
-    @DisplayName("자동차 경주 게임 한 라운드 진행")
-    @Test
-    void play_one_round() {
-        CarRacingGame carRacingGame = new CarRacingGame(
-                "pobi,crong,honux",
-                5,
-                new TestNumberGenerator(
-                        List.of(0, 2, 1, 8, 5, 4, 9, 6, 7, 3, 1, 5, 1, 1, 7, 3, 6, 9, 4, 4, 5, 1, 2, 2, 6)
-                )
-        );
-        assertThat(carRacingGame.playOneRound()).containsExactly("pobi : -", "crong : -", "honux : -");
-    }
-
     @DisplayName("자동차 경주 게임 여러 라운드 진행")
     @Test
     void play_several_rounds() {
